Match absolute patterns without volume on windows
In 1.1.5 Glob() on windows matched absolute patterns without volume specified with the current working dir's volume.
This broke in 1.2.x due to the use of filepath.IsAbs() which returns false if the volume is not present.
This was causing paths starting with / or \ on windows to be erroneously evaluated from the current directory.