add option to fail on I/O errors
This commit introduces an Option for Glob(), GlobWalk() and
FilepathGlob() to handle encountered I/O errors as fatal.
When the option is enabled and an I/O error happens, globbing is aborted
and the error is returned.
By default the globbing function continue when an I/O error happens and
only skip the related file an directory. This is the same behaviour as
before.

The new option is useful when the user should be notified about broken
symlinks, non-existing paths in the pattern or general I/O errors that
were encountered.

An Opt type is introduced that follows the Option pattern.
Opt is returned by the WithFailOnIOErrors() function.
The exported functions are changed to accept as last argument a variadic
...Opt argument.
When the functions are called, they instantiate a glob type instance,
that stores the value of the option in an failOnIOErrors field.
All functions that do I/O operations are converted to being methods of
glob, to be able to access failOnIOErrors.
If an I/O operation encounters an error and failOnIOErrors is set, the
error is returned.
If failOnIOErrors is not set, the functions behave the same as before.

The introduced Opt type and glob struct allow to extend the package with
further options, without having to break the API in the future.

Testcases were added for Glob and GlobWalk to check if
WithFailOnIOErrors causes these functions to abort and return an error.

The code could be refactored in a follow-up to use the new glob type to
store the arguments that are currently passed around on functions calls
like fsys and pattern.